+++ services/app-api/package.json
@@ -13,6 +13,7 @@
     "url": "https://github.com/AnomalyInnovations/serverless-nodejs-starter.git"
   },
   "devDependencies": {
+    "@types/uuid": "^8.3.4",
     "serverless-api-client-certificate": "1.0.2",
     "serverless-associate-waf": "^1.2.1"
   },
